Refactoring (#136)
* tabs and spaces

* tabs, spaces

* model Settings

* add delete() for Settings

* move getSiteMenu

* tabs

* remove function

* don't show subpages for news

* codex.special update

* move sidebar block

* update for site_info

using Model_Settings

* tabs, spaces

* tabs, spaces

* no need

* migrations

* tabs, spaces

* upd

* edit database.sample

* add _.htaccess

* open photo in new tab

* no need

* delete /admin section

* upd contacts. map

* upd for main.css

* upd for js-comments

* upd for contacts

* upd

* upd

* upd htaccess

* remove files from form while creating a page

* delete and rename files

* upd

* add cache

* upd

* upd codex-special

* add fresh migrations. remove the rest

* улучшеное is_removed для страницы

при удалении страницы удаляются все файлы, комментарии и подстраницы.
у всего вышеперечисленног is_removed будет равен 1

* add js-approval-button

* reuse navigation bar

* site_menu cache

* missing $attachments

* add is_menu_item filter

* upd

* upd for from

* upd

* upd

* log errors

* hidden "add menu item"

* upd

+++ application/classes/Controller/Comments.php
@@ -2,10 +2,8 @@
 
 class Controller_Comments extends Controller_Base_preDispatch
 {
-
     public function action_add()
     {
-
         $comment = new Model_Comment();
 
         $page = new Model_Page($this->request->param('id'));
@@ -17,17 +15,29 @@ class Controller_Comments extends Controller_Base_preDispatch
         /**
          * Checking for existing page
          */
-        if (!$page->id) { $error = 'Wrong page id'; goto finish; }
+        if (!$page->id) {
+
+            $error = 'Wrong page id';
+            goto finish;
+        }
 
         /**
          * Checking for authorized user
          */
-        if ($this->user->status < Model_User::USER_STATUS_REGISTERED) { $error = 'Access denied'; goto finish; }
+        if ($this->user->status < Model_User::USER_STATUS_REGISTERED) {
+
+            $error = 'Access denied';
+            goto finish;
+        }
 
         /**
          * Checking for existing text
          */
-        if (!$text) { $error = 'Text is incorrect'; goto finish; }
+        if (!$text) {
+
+            $error = 'Text is incorrect';
+            goto finish;
+        }
 
         $comment->page_id              = $this->request->param('id');
         $comment->text                 = $text;
@@ -39,8 +49,11 @@ class Controller_Comments extends Controller_Base_preDispatch
 
         finish:
         if ($error) {
+
             $this->redirect( '/p/' . $page->id . '/' . $page->uri . '?error=' . $error );
+
         } else {
+
             $this->redirect( '/p/' . $page->id . '/' . $page->uri );
         }
     }
@@ -51,14 +64,10 @@ class Controller_Comments extends Controller_Base_preDispatch
 
         $comment = Model_Comment::get($comment_id);
 
-        if ($comment->author->id == $this->user->id || $this->user->isAdmin)
-        {
-            $comment->delete();
-        }
+        if ($comment->author->id == $this->user->id || $this->user->isAdmin) $comment->delete();
 
         $page = new Model_Page($comment->page_id);
 
         $this->redirect('/p/' . $page->id . '/' . $page->uri);
     }
-
 }
